feat: block buffer tree
添加缓冲区结构来进行区块分叉的处理

feat: construct with buffer

fixed: Check genesis block

修复了对创世区块判断的bug

fixed: set type uint64 -> int64

added: dir gitkeep

feat: syncer msg handler

test: buffer test pass

+++ core/test_tools.go
@@ -11,6 +11,7 @@ import (
 	"time"
 )
 
+// buildTransaction 仅用于测试，也是一个构建交易的例子
 func buildTransaction(key *ecdsa.PrivateKey) *common.Transaction {
 	data := make([]byte, 32)
 	rand.Read(data)
@@ -18,8 +19,8 @@ func buildTransaction(key *ecdsa.PrivateKey) *common.Transaction {
 
 	txBody := common.TransactionBody{
 		Data:      data,
-		Timestamp: uint64(timestamp),
-		Expire:    uint64(timestamp + 3000),
+		Timestamp: timestamp,
+		Expire:    timestamp + 3000,
 	}
 
 	txBody.Public = [33]byte(crypto.PublicKey2Bytes(&key.PublicKey))
