+++ utils/serializer_test.go
@@ -20,8 +20,8 @@ func buildTransaction(key *ecdsa.PrivateKey) *common.Transaction {
 
 	txBody := common.TransactionBody{
 		Data:      data,
-		Timestamp: uint64(timestamp),
-		Expire:    uint64(timestamp + 3000),
+		Timestamp: timestamp,
+		Expire:    timestamp + 3000,
 	}
 
 	txBody.Public = [33]byte(crypto.PublicKey2Bytes(&key.PublicKey))
